feat: use vue components are shaperone property/object renderers
This will give us more control over the styling of our forms.

One downside I still couldn't figure out is that there are a lot of
superfluous HTML tags in the DOM, which makes styling way harder than it
should be. But at least we can have a consistent look and change the
rendering as needed.